requirements.txt
-e git+https://github.com/Runnerly/flakon.git#egg=flakon
python-decouple
requests